requirements.txt
fastapi>=0.103.1,<0.104

#web server for fastapi
uvicorn>=0.20.0,<0.21

# ORM for fastapi
SQLAlchemy>=2.0.7,<2.1

# for hashing passwords
passlib>=1.7.4,<1.8
bcrypt>=4.0.1,<4.1

# for outh2 password request forms
python-multipart>=0.0.6,<0.1

# for jwt tokens
python-jose>=3.3.0,<3.4
cryptography>=39.0.2,<39.1

# for postgres
psycopg2-binary>=2.9.6,<2.10
